byol_pytorch/requirements.txt
absl-py==2.1.0
accelerate==0.30.1
aiohttp==3.9.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
apturl==0.5.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
bcrypt==3.2.0
beautifulsoup4==4.12.3
blinker==1.4
boto3==1.34.118
botocore==1.34.118
bpemb==0.3.5
Brlapi==0.8.3
cachetools==5.3.3
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
conllu==4.5.3
contourpy==1.2.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
datasets==2.19.2
dbus-python==1.2.18
debugpy==1.8.1
decorator==5.1.1
defer==1.0.6
Deprecated==1.2.14
dill==0.3.8
diskcache==5.6.3
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
exceptiongroup==1.2.1
executing==2.0.1
fasteners==0.14.1
filelock==3.13.4
flair==0.13.1
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
future==0.18.2
gdown==5.2.0
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.43
gliner==0.2.2
grpcio==1.62.2
h5py==3.11.0
httplib2==0.20.2
huggingface-hub==0.23.2
hydra-core==1.3.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.29.4
ipython==8.23.0
Janome==0.5.0
jax==0.4.28
jaxlib==0.4.28
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.3.3
keyring==23.5.0
kiwisolver==1.4.5
langdetect==1.0.9
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
librosa==0.10.2.post1
llama_cpp_python==0.2.77
llvmlite==0.42.0
lockfile==0.12.2
louis==3.20.0
lxml==5.2.2
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.14
ml-dtypes==0.4.0
monotonic==1.6
more-itertools==10.2.0
mpld3==0.5.10
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
namex==0.0.8
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.133
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.2
oauthlib==3.2.0
olefile==0.46
omegaconf==2.3.0
opencv-contrib-python==4.10.0.82
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
paramiko==2.9.3
parso==0.8.4
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.1
pooch==1.8.1
pptree==3.1
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
pytorch_revgrad==0.2.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.0.3
regex==2024.5.15
reportlab==3.6.8
requests==2.32.3
rich==13.7.1
s3transfer==0.10.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.12.0
screen-resolution-extra==0.0.0
seaborn==0.13.2
SecretStorage==3.3.1
segtok==1.5.11
semver==3.0.2
sentencepiece==0.2.0
seqeval==1.2.2
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sounddevice==0.4.7
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
sqlitedict==2.1.0
ssh-import-id==5.11
stack-data==0.6.3
sympy==1.12
systemd-python==234
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.2.0
torchaudio==2.2.2
torchvision==0.17.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
transformer-smaller-training-vocab==0.4.0
transformers==4.41.2
triton==2.2.0
typing_extensions==4.12.1
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
ultralytics==8.2.11
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.13.0+ds
wadllib==1.3.6
wcwidth==0.2.13
Werkzeug==3.0.2
Wikipedia-API==0.6.0
wrapt==1.16.0
xdg==5
xkit==0.0.0
xxhash==3.4.1
yarl==1.9.4
zipp==1.0.0
